+++ R/mc_metadata.R
@@ -3,21 +3,18 @@
 #' Access the Multi-CAST metadata
 #'
 #' \code{mc_metadata} downloads a table with metadata on the texts and speakers
-#' in the Multi-CAST collection. The data is downloaded from the servers of
-#' University of Bamberg and presented as a
-#' \code{\link[data.table]{data.table}}.
+#' in the Multi-CAST collection from the servers of the University of Bamberg.
 #'
 #' @seealso \code{\link{multicast}}, \code{\link{mc_index}},
 #'   \code{\link{mc_referents}}, \code{\link{mc_clauses}}
 #'
-#' @param vkey A numeric or character vector of length 1 specifying the
-#'   requested version of the annotation values. Must be one of the four-digit
-#'   version keys in the first column of \code{\link{mc_index}}, or empty. If
-#'   empty or no value is supplied, the most recent version of the annotations
-#'   is retrieved automatically.
+#' @param vkey A four-digit number specifying the requested version of the
+#'   metadata. Must be one of the version keys listed in the first column of
+#'   \code{\link{mc_index}}, or empty. If empty, the most recent version of the
+#'   metadata is retrieved automatically.
 #'
-#' @return A \code{\link[data.table]{data.table}} containing metadata on the
-#'   Multi-CAST collection. The table has the following eight columns:
+#' @return A \code{\link{data.frame}} containing metadata on the Multi-CAST
+#'   collection. The table has the following eight columns:
 #'
 #'   \describe{ \item{\code{[, 1] corpus}}{The name of the corpus.}
 #'   \item{\code{[, 2] text}}{The title of the text.} \item{\code{[, 3]
@@ -35,103 +32,78 @@
 #'   # retrieve the most recent version of the Multi-CAST metadata
 #'   mc_metadata()
 #'
-#'   # retrieve the lists of referents published in May 2019
-#'   mc_metadata(1905)   # or: mc_metadata("1905")
+#'   # retrieve the lists of referents published in January 2021
+#'   mc_metadata(2101)
 #'
 #'   # join the metadata to a table with annotation values
 #'   mc <- multicast()
-#'   merge(mc, mc_metadata(), by = c("corpus", "text"))
+#'   merge(mc, mc_metadata(),
+#'         by = c("corpus", "text"))
 #' }
 #'
-#' @importFrom curl curl
 #' @export
-mc_metadata <- function(vkey) {
-	# check whether vkey is missing
-	if (!mc_missarg(vkey)) {
-		# 1A: vkey is not missing
-		# check whether vkey is a numeric or character vector
-		if (!(is.numeric(vkey) | is.character(vkey))) {
-			# 2A: vkey is not a numeric or character vector
-			stop(paste0("\n  Argument is not of type numeric or character."))
-		} else {
-			# 2B: vkey has correct type
-			# check whether vkey has length == 1
-			if (length(vkey) != 1) {
-				# 3A: vkey has length != 1
-				stop(paste0("\n  Argument has length > 1."))
+mc_metadata <- function(vkey = NULL) {
+
+	# fetch version index
+	index <- mc_index()
+	useDefaultKey <- FALSE
+
+	# check for invalid input
+	if (!(is.null(vkey))) {
+		if (length(vkey) > 1) {
+			stop("Invalid version key. Argument has length > 1.")
+		} else if (!(is.na(vkey) | vkey == "")) {
+			if (!(is.numeric(vkey) | is.character(vkey))) {
+				stop(paste0("Invalid version key. Argument must be numeric or character."))
 			} else {
-				# 3B: vkey has length == 1
-				# check whether vkey has valid format
-				if (!grepl("^\\d\\d\\d\\d$", vkey)) {
-					# 4A: vkey has invalid format
-					stop(paste0("\n  The supplied version key '",
-								vkey,
-								"' has invalid format."))
-				} else {
-					# 4B: vkey has valid format
-
-					# fetch version index
-					index <- mc_index()
-
-					# check whether vkey is in index
-					if (!any(index[, 1] == vkey)) {
-						# 5A: vkey is not in index
-						stop(paste0("\n  The requested version '",
-									vkey,
-									"' does not exist."))
-					} else {
-						# 5B: vkey found
-
-						# add 'latest' tag to message
-						if (vkey==index[1, 1]) {
-							lat <- "latest, "
-						} else {
-							lat <- ""
-						}
-						message(paste0("Found requested version '",
-									   vkey,
-									   "' (",
-									   lat,
-									   "published ",
-									   index[version == vkey,2], ")."))
-					}
+				if (!(vkey %in% index[[1]])) {
+					stop("Invalid version key. Specified version not found.")
 				}
 			}
+		} else {
+			useDefaultKey <- TRUE
 		}
 	} else {
-		# 1B: vkey is missing
+		useDefaultKey <- TRUE
+	}
 
-		# fetch version index
-		index <- mc_index()
-
-		# select latest vkey from index
-		vkey <- as.character(index[1, 1])
-		message(paste0("Found latest version '",
+	# if no index was supplied, pick the most recent
+	if (useDefaultKey) {
+		vkey <- index[1, 1]
+		message(paste0("No version specified, defaulting to latest ('",
 					   index[1, 1],
-					   "' (published ",
+					   "', published ",
 					   index[1, 2], ")."))
 	}
 
-	# construct URL for annotation file
+
+	# download the data
 	path <- paste0("https://multicast.aspra.uni-bamberg.de/data/docs/general/metadata/",
 				   vkey,
 				   "__mc_metadata.tsv")
 
-	# fetch annotation file
-	message("Retrieving metadata...")
-	tryCatch(
-		suppressWarnings(
-			meta <- data.table::fread(path,
-									  sep = "\t",
-									  header = TRUE,
-									  colClasses = list(factor = c(1:3, 5:6), character = c(4, 7:8)),
-									  encoding = "UTF-8",
-									  showProgress = FALSE)
-		),
-		error = function(e) { stop("failed to download metadata.") }
-	)
-#	message(paste0("Success!"))
+	message("Retrieving Multi-CAST metadata...")
+	meta <- tryCatch(suppressWarnings(read.csv(path,
+											   sep = "\t",
+											   header = TRUE,
+											   colClasses = c(rep("factor", 3),
+											   			      "character",
+											   			      rep("factor", 2),
+											   			      rep("character", 2)),
+											   encoding = "UTF-8")),
+				   error = function(e) {
+				   		      stop(paste0("Failed to download data. Cannot access file.\n",
+					  		     	      "       The servers of the University of Bamberg seem to be experiencing problems.\n",
+										  "       Please try again later."),
+								   call. = FALSE)
+					  	   }
+			       )
+	message(paste0("Downloaded <1 MB."))
 
-	# return annotation values
+	# return table with annotation values
 	return(meta)
+
 }
+
+
+# ----------------------------------------------------------------------