+++ R/mc_referents.R
@@ -2,26 +2,25 @@
 
 #' Access the Multi-CAST list of referents
 #'
-#' \code{mc_referents} downloads the lists of referents for all texts in the
-#' Multi-CAST collection that have been annotated with the RefIND scheme
-#' (Referent Indexing in Natural-language Discourse, Schiborr et al. 2018). The
-#' data are downloaded from the servers of University of Bamberg and presented as
-#' a \code{\link[data.table]{data.table}}.
+#' \code{mc_referents} downloads a tabular list of all discourse referents
+#' occuring in those texts in the Multi-CAST collection that have been annotated
+#' with the RefIND scheme (Schiborr et al. 2018). The data are downloaded from
+#' the servers of University of Bamberg.
 #'
 #' @seealso \code{\link{multicast}}, \code{\link{mc_index}},
 #'   \code{\link{mc_metadata}}, \code{\link{mc_clauses}}
 #'
-#' @param vkey A numeric or character vector of length 1 specifying the
-#'   requested version of the annotation values. Must be one of the four-digit
-#'   version keys in the first column of \code{\link{mc_index}}, or empty. If
-#'   empty or no value is supplied, the most recent version of the annotations
-#'   is retrieved automatically. Note that the first annotations with RefIND
-#'   were added with version 1905 (May 2019), and hence no lists of referents
-#'   exist for earlier versions (i.e. 1505 and 1606).
+#' @param vkey A four-digit number specifying the requested version of the list
+#'   of referents. Must be one of the version keys listed in the first column of
+#'   \code{\link{mc_index}}, or empty. If empty, the most recent version of the
+#'   list of referents is retrieved automatically. Note that the first
+#'   annotations with RefIND were added with version 1905 (May 2019) of
+#'   Multi-CAST, and hence no lists of referents exist for earlier versions
+#'   (i.e. 1505 and 1606).
 #'
-#' @return A \code{\link[data.table]{data.table}} containing lists of referents
-#'   for all texts with RefIND annotations in the Multi-CAST collection. The table
-#'   has the following eight columns:
+#' @return A \code{\link{data.frame}} containing a list of referents for all
+#'   texts with RefIND annotations in the Multi-CAST collection. The table has
+#'   the following eight columns:
 #'
 #'   \describe{ \item{\code{[, 1] corpus}}{The name of the corpus.}
 #'   \item{\code{[, 2] text}}{The title of the text.} \item{\code{[, 3]
@@ -39,11 +38,11 @@
 #'
 #' @examples
 #' \dontrun{
-#'   # retrieve the most recent version of the Multi-CAST lists of referents
+#'   # retrieve the most recent version of the Multi-CAST list of referents
 #'   mc_referents()
 #'
-#'   # retrieve the lists of referents published in May 2019
-#'   mc_referents(1905)   # or: mc_referents("1905")
+#'   # retrieve the lists of referents published in January 2021
+#'   mc_referents(2021)
 #'
 #'   # join the list of referents to a table with annotation values
 #'   mc <- multicast()
@@ -52,103 +51,70 @@
 #'         all.x = TRUE)
 #' }
 #'
-#' @importFrom curl curl
 #' @export
-mc_referents <- function(vkey) {
-	# check whether vkey is missing
-	if (!mc_missarg(vkey)) {
-		# 1A: vkey is not missing
-		# check whether vkey is a numeric or character vector
-		if (!(is.numeric(vkey) | is.character(vkey))) {
-			# 2A: vkey is not a numeric or character vector
-			stop(paste0("\n  Argument is not of type numeric or character."))
-		} else {
-			# 2B: vkey has correct type
-			# check whether vkey has length == 1
-			if (length(vkey) != 1) {
-				# 3A: vkey has length != 1
-				stop(paste0("\n  Argument has length > 1."))
+mc_referents <- function(vkey = NULL) {
+
+	# fetch version index
+	index <- mc_index()
+	useDefaultKey <- FALSE
+
+	# check for invalid input
+	if (!(is.null(vkey))) {
+		if (length(vkey) > 1) {
+			stop("Invalid version key. Argument has length > 1.")
+		} else if (!(is.na(vkey) | vkey == "")) {
+			if (!(is.numeric(vkey) | is.character(vkey))) {
+				stop(paste0("Invalid version key. Argument must be numeric or character."))
 			} else {
-				# 3B: vkey has length == 1
-				# check whether vkey has valid format
-				if (!grepl("^\\d\\d\\d\\d$", vkey)) {
-					# 4A: vkey has invalid format
-					stop(paste0("\n  The supplied version key '",
-								vkey,
-								"' has invalid format."))
-				} else {
-					# 4B: vkey has valid format
-
-					# fetch version index
-					index <- mc_index()
-
-					# check whether vkey is in index
-					if (!any(index[, 1] == vkey)) {
-						# 5A: vkey is not in index
-						stop(paste0("\n  The requested version '",
-									vkey,
-									"' does not exist."))
-					} else {
-						# 5B: vkey found
-
-						# check whether vkey is for version with no RefIND
-						if (vkey == "1505" | vkey == "1606") {
-							# 6A: version has no RefIND
-							stop(paste0("\n No lists of referents exist before version 1905 of Multi-CAST."))
-						} else {
-							# 6B: version has RefIND
-
-							# add 'latest' tag to message
-							if (vkey == index[1, 1]) {
-								lat <- "latest, "
-							} else {
-								lat <- ""
-							}
-							message(paste0("Found requested version '",
-										   vkey,
-										   "' (",
-										   lat,
-										   "published ",
-										   index[version == vkey,2], ")."))
-						}
-					}
+				if (vkey == "1505" | vkey == "1606") {
+					stop("No lists of referents exist for versions '1505' and '1606' of Multi-CAST.")
+				}
+				else if (!(vkey %in% index[[1]])) {
+					stop("Invalid version key. Specified version not found.")
 				}
 			}
+		} else {
+			useDefaultKey <- TRUE
 		}
 	} else {
-		# 1B: vkey is missing
+		useDefaultKey <- TRUE
+	}
 
-		# fetch version index
-		index <- mc_index()
-
-		# select latest vkey from index
-		vkey <- as.character(index[1, 1])
-		message(paste0("Found latest version '",
+	# if no index was supplied, pick the most recent
+	if (useDefaultKey) {
+		vkey <- index[1, 1]
+		message(paste0("No version specified, defaulting to latest ('",
 					   index[1, 1],
-					   "' (published ",
+					   "', published ",
 					   index[1, 2], ")."))
 	}
 
-	# construct URL for annotation file
+
+	# download the data
 	path <- paste0("https://multicast.aspra.uni-bamberg.de/data/docs/general/list-of-referents/",
 				   vkey,
 				   "__mc_list-of-referents.tsv")
 
-	# fetch annotation file
-	message("Retrieving list of referents...")
-	tryCatch(
-		suppressWarnings(
-			refs <- data.table::fread(path,
-									  sep = "\t",
-									  header = TRUE,
-									  colClasses = list(factor = 1:2, character = 3:8),
-									  encoding = "UTF-8",
-									  showProgress = FALSE)
-		),
-		error = function(e) { stop("failed to download list of referents.") }
-	)
-#	message(paste0("Success!"))
+	message("Retrieving Multi-CAST list of referents...")
+	meta <- tryCatch(suppressWarnings(read.csv(path,
+											   sep = "\t",
+											   header = TRUE,
+											   colClasses = c(rep("factor", 2),
+											   			      rep("character", 5)),
+											   encoding = "UTF-8")),
+				   error = function(e) {
+				   		      stop(paste0("Failed to download data. Cannot access file.\n",
+					  		     	      "       The servers of the University of Bamberg seem to be experiencing problems.\n",
+										  "       Please try again later."),
+								   call. = FALSE)
+					  	   }
+			       )
+	message(paste0("Downloaded <1 MB."))
 
-	# return annotation values
-	return(refs)
+	# return table with annotation values
+	return(meta)
+
 }
+
+
+# ----------------------------------------------------------------------